requirements.txt
psycopg[binary,pool]
sqlalchemy
pandas
python-dotenv
rich
openai
streamlit
